pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "tsp"
version = "0.0.1"
description = "Given features predict Survival"
authors = [
  { name = "Mrudhuhas" },
]

readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    
]
dependencies = [
    "boto3>=1.39.12",
    "google-cloud-storage>=3.2.0",
    "great-expectations>=1.5.5",
    "hydra-core>=1.3.2",
    "loguru",
    "mkdocs",
    "mlflow>=3.1.4",
    "pandera>=0.25.0",
    "pip",
    "prefect>=3.4.10",
    "pytest",
    "python-dotenv",
    "ruff",
    "tqdm",
    "typer",
]
requires-python = "~=3.10.0"


[tool.ruff]
line-length = 99
src = ["tsp"]
include = ["pyproject.toml", "tsp/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["tsp"]
force-sort-within-sections = true

[dependency-groups]
dev = [
    "ipykernel>=6.30.0",
    "kagglehub>=0.3.12",
]

